move RR from ./index to ./rr
necessary because rr/* inherit from it, and loading ../index from rr/* caused a circular dependency failure
