+++ rr/aaaa.js
@@ -1,7 +1,7 @@
 
 import net from 'net'
 
-import RR from '../index.js'
+import RR from '../rr.js'
 import * as TINYDNS from '../lib/tinydns.js'
 
 export default class AAAA extends RR {
