+++ rr/dname.js
@@ -1,6 +1,6 @@
 import net from 'net'
 
-import RR from '../index.js'
+import RR from '../rr.js'
 import * as TINYDNS from '../lib/tinydns.js'
 
 export default class DNAME extends RR {
