+++ rr/naptr.js
@@ -1,5 +1,5 @@
 
-import RR from '../index.js'
+import RR from '../rr.js'
 import * as TINYDNS from '../lib/tinydns.js'
 
 const rdataRe = /[\r\n\t:\\/]/
