+++ rr/nsec3.js
@@ -1,5 +1,5 @@
 
-import RR from '../index.js'
+import RR from '../rr.js'
 
 export default class NSEC3 extends RR {
   constructor (opts) {
