+++ rr/nsec3param.js
@@ -1,5 +1,5 @@
 
-import RR from '../index.js'
+import RR from '../rr.js'
 
 export default class NSEC3PARAM extends RR {
   constructor (opts) {
